+++ src/app/portfolio/[slug]/page.tsx
@@ -125,7 +125,10 @@ export default async function PortfolioGalleryPage({
       />
       <SiteHeader />
       {theme && cover && (
-        <div className="relative aspect-[16/9] w-full overflow-hidden sm:aspect-[21/9]">
+        <div
+          className="relative w-full overflow-hidden"
+          style={{ height: 'var(--gallery-cover-height)' }}
+        >
           {/* eslint-disable-next-line @next/next/no-img-element */}
           <img
             src={`/img/${cover}/web`}